chore: Refactor logging, version and config
The three modules have been migrated to a core package.
