+++ src/pdm_bump/core/__init__.py
@@ -1,0 +1,11 @@
+#
+# SPDX-License-Identifier: MIT
+#
+# Copyright (c) 2021-2023 Carsten Igel.
+#
+# This file is part of pdm-bump
+# (see https://github.com/carstencodes/pdm-bump).
+#
+# This file is published using the MIT license.
+# Refer to LICENSE for more information
+#
