+++ src/pdm_bump/vcs/gitcli.py
@@ -14,8 +14,8 @@ from pathlib import Path
 from subprocess import CalledProcessError
 from typing import Optional, cast
 
-from ..logging import logger
-from ..version import Version
+from ..core.logging import logger
+from ..core.version import Version
 from .core import VcsProvider, VcsProviderError, vcs_provider
 from .git import GitCommonVcsProviderFactory
 from .mixins import CliRunnerMixin
